updated with playwright fix
